third_party/WebKit/LayoutTests/shadow-dom/imperative-api.html

<!DOCTYPE html>
<!--
Tests for Imperative Shadow DOM Distribution API.
See https://crbug.com/869308
-->
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="host">
  <div id="child1"></div>
</div>

<script>
const host = document.querySelector('#host');
const child1 = document.querySelector('#child1');
const shadow_root = host.attachShadow({ mode: 'open', slotting: 'manual' });
const slot1 = document.createElement('slot');
shadow_root.appendChild(slot1);

test(() => {
  assert_equals(child1.assignedSlot, null);
  assert_array_equals(slot1.assignedElements(), []);

  slot1.assign([child1]);

  assert_equals(child1.assignedSlot, slot1);
  assert_array_equals(slot1.assignedElements(), [child1]);
}, 'slot.assign(nodes) should work');
</script>
